remove ExplainerHeader but use PosLabelConnector
