+++ explainerdashboard/dashboard_components/composites.py
@@ -22,7 +22,7 @@ from .decisiontree_components import *
 class ClassifierModelStatsComposite(ExplainerComponent):
     def __init__(self, explainer, title="Classification Stats", 
                     header_mode="none", name=None,
-                    hide_selector=False,
+                    hide_selector=True,
                     bin_size=0.1, quantiles=10, cutoff=0.5):
         """Composite of multiple classifier related components: 
             - precision graph
@@ -42,6 +42,7 @@ class ClassifierModelStatsComposite(ExplainerComponent):
             name (str, optional): unique name to add to Component elements. 
                         If None then random uuid is generated to make sure 
                         it's unique. Defaults to None.
+            hide_selector(bool, optional): hide all pos label selectors. Defaults to True.
             bin_size (float, optional): bin_size for precision plot. Defaults to 0.1.
             quantiles (int, optional): number of quantiles for precision plot. Defaults to 10.
             cutoff (float, optional): initial cutoff. Defaults to 0.5.
@@ -55,7 +56,7 @@ class ClassifierModelStatsComposite(ExplainerComponent):
         self.rocauc = RocAucComponent(explainer, hide_selector=hide_selector)
         self.prauc = PrAucComponent(explainer, hide_selector=hide_selector)
 
-        self.cutoffpercentile = CutoffPercentileComponent(explainer)
+        self.cutoffpercentile = CutoffPercentileComponent(explainer, hide_selector=hide_selector)
         self.cutoffconnector = CutoffConnector(self.cutoffpercentile,
                 [self.precision, self.confusionmatrix, self.liftcurve, 
                  self.classification, self.rocauc, self.prauc])
@@ -165,7 +166,8 @@ class RegressionModelStatsComposite(ExplainerComponent):
 
 class IndividualPredictionsComposite(ExplainerComponent):
     def __init__(self, explainer, title="Individual Predictions",
-                        header_mode="none", name=None):
+                        header_mode="none", name=None,
+                        hide_selector=True):
         """Composite for a number of component that deal with individual predictions:
 
         - random index selector
@@ -184,17 +186,23 @@ class IndividualPredictionsComposite(ExplainerComponent):
             name (str, optional): unique name to add to Component elements. 
                         If None then random uuid is generated to make sure 
                         it's unique. Defaults to None.
+            hide_selector(bool, optional): hide all pos label selectors. Defaults to True.
         """
         super().__init__(explainer, title, header_mode, name)
 
         if self.explainer.is_classifier:
-            self.index = ClassifierRandomIndexComponent(explainer)
+            self.index = ClassifierRandomIndexComponent(
+                            explainer, hide_selector=hide_selector)
         elif self.explainer.is_regression:
             self.index = RegressionRandomIndexComponent(explainer)
-        self.summary = PredictionSummaryComponent(explainer)
-        self.contributions = ShapContributionsGraphComponent(explainer)
-        self.pdp = PdpComponent(explainer)
-        self.contributions_list = ShapContributionsTableComponent(explainer)
+        self.summary = PredictionSummaryComponent(
+                            explainer, hide_selector=hide_selector)
+        self.contributions = ShapContributionsGraphComponent(
+                                explainer, hide_selector=hide_selector)
+        self.pdp = PdpComponent(
+                        explainer, hide_selector=hide_selector)
+        self.contributions_list = ShapContributionsTableComponent(
+                                        explainer, hide_selector=hide_selector)
 
         self.index_connector = IndexConnector(self.index, 
                 [self.summary, self.contributions, self.pdp, self.contributions_list])
@@ -235,6 +243,7 @@ class IndividualPredictionsComposite(ExplainerComponent):
 class ShapDependenceComposite(ExplainerComponent):
     def __init__(self, explainer, title='Feature Dependence',
                     header_mode="none", name=None,
+                    hide_selector=True,
                     depth=None, cats=True):
         """Composite of ShapSummary and ShapDependence component
 
@@ -248,15 +257,22 @@ class ShapDependenceComposite(ExplainerComponent):
             name (str, optional): unique name to add to Component elements. 
                         If None then random uuid is generated to make sure 
                         it's unique. Defaults to None.
+            hide_selector (bool, optional): hide all pos label selectors. Defaults to True.
             depth (int, optional): Number of features to display. Defaults to None.
             cats (bool, optional): Group categorical features. Defaults to True.
         """
         super().__init__(explainer, title, header_mode, name)
         
-        self.shap_summary = ShapSummaryComponent(self.explainer, depth=depth, cats=cats)
+        self.shap_summary = ShapSummaryComponent(
+                    self.explainer, 
+                    hide_selector=hide_selector, 
+                    depth=depth, cats=cats)
         self.shap_dependence = ShapDependenceComponent(
-                                    self.explainer, hide_cats=True, cats=cats)
-        self.connector = ShapSummaryDependenceConnector(self.shap_summary, self.shap_dependence)
+                    self.explainer, 
+                    hide_selector=hide_selector, hide_cats=True, 
+                    cats=cats)
+        self.connector = ShapSummaryDependenceConnector(
+                    self.shap_summary, self.shap_dependence)
         self.register_components(self.shap_summary, self.shap_dependence, self.connector)
 
     def _layout(self):
@@ -275,6 +291,7 @@ class ShapDependenceComposite(ExplainerComponent):
 class ShapInteractionsComposite(ExplainerComponent):
     def __init__(self, explainer, title='Feature Interactions',
                     header_mode="none", name=None,
+                    hide_selector=True,
                     depth=None, cats=True):
         """Composite of InteractionSummaryComponent and InteractionDependenceComponent
 
@@ -288,15 +305,16 @@ class ShapInteractionsComposite(ExplainerComponent):
             name (str, optional): unique name to add to Component elements. 
                         If None then random uuid is generated to make sure 
                         it's unique. Defaults to None.
+            hide_selector (bool, optional): hide all pos label selectors. Defaults to True.
             depth (int, optional): Initial number of features to display. Defaults to None.
             cats (bool, optional): Initally group cats. Defaults to True.
         """
         super().__init__(explainer, title, header_mode, name)
 
         self.interaction_summary = InteractionSummaryComponent(
-                explainer, depth=depth, cats=cats)
+                explainer, hide_selector=hide_selector, depth=depth, cats=cats)
         self.interaction_dependence = InteractionDependenceComponent(
-                explainer, cats=cats)
+                explainer, hide_selector=hide_selector, cats=cats)
         self.connector = InteractionSummaryDependenceConnector(
             self.interaction_summary, self.interaction_dependence)
         self.register_components(
@@ -304,20 +322,23 @@ class ShapInteractionsComposite(ExplainerComponent):
         
     def _layout(self):
         return html.Div([
-            dbc.Row([
-                dbc.Col([
-                    self.interaction_summary.layout()
-                ], width=6),
-                dbc.Col([
-                    self.interaction_dependence.layout()
-                ], width=6),
-            ])
+            dbc.Container([
+                dbc.Row([
+                    dbc.Col([
+                        self.interaction_summary.layout()
+                    ], width=6),
+                    dbc.Col([
+                        self.interaction_dependence.layout()
+                    ], width=6),
+                ])
+            ], fluid=True)  
         ])
 
 
 class DecisionTreesComposite(ExplainerComponent):
     def __init__(self, explainer, title="Decision Trees",
-                    header_mode="none", name=None):
+                    header_mode="none", name=None,
+                    hide_selector=True):
         """Composite of decision tree related components:
         
         - individual decision trees barchart
@@ -334,13 +355,14 @@ class DecisionTreesComposite(ExplainerComponent):
             name (str, optional): unique name to add to Component elements. 
                         If None then random uuid is generated to make sure 
                         it's unique. Defaults to None.
+            hide_selector (bool, optional): hide all pos label selectors. Defaults to True.
         """
         super().__init__(explainer, title, header_mode, name)
         
-        self.index = ClassifierRandomIndexComponent(explainer)
-        self.trees = DecisionTreesComponent(explainer)
-        self.decisionpath_table = DecisionPathTableComponent(explainer)
-        self.decisionpath_graph = DecisionPathGraphComponent(explainer)
+        self.index = ClassifierRandomIndexComponent(explainer, hide_selector=hide_selector)
+        self.trees = DecisionTreesComponent(explainer, hide_selector=hide_selector)
+        self.decisionpath_table = DecisionPathTableComponent(explainer, hide_selector=hide_selector)
+        self.decisionpath_graph = DecisionPathGraphComponent(explainer, hide_selector=hide_selector)
 
         self.index_connector = IndexConnector(self.index, 
             [self.trees, self.decisionpath_table, self.decisionpath_graph])
