+++ explainerdashboard/dashboard_components/dashboard_methods.py
@@ -6,8 +6,6 @@ __all__ = [
     'ExplainerComponent',
     'PosLabelSelector',
     'make_hideable',
-    'ExplainerTabsLayout',
-    'ExplainerPageLayout'
 ]
 
 from abc import ABC
@@ -102,104 +100,104 @@ class DummyComponent:
 
 
 
-class ExplainerHeader:
-    """
-    Generates a header layout with a title and, for classification models, a
-    positive label selector. The callbacks for most of the ExplainerComponents
-    expect the presence of both a 'pos-label' and a 'tabs' dash element, so you
-    have to make sure there is at least an element with that name.
+# class ExplainerHeader:
+#     """
+#     Generates a header layout with a title and, for classification models, a
+#     positive label selector. The callbacks for most of the ExplainerComponents
+#     expect the presence of both a 'pos-label' and a 'tabs' dash element, so you
+#     have to make sure there is at least an element with that name.
 
-    If you have standalone page without tabs ans so without a dcc.Tabs 
-    component, an ExplainerHeader(mode='standalone') will insert a dummy 
-    'tabs' hidden div.
+#     If you have standalone page without tabs ans so without a dcc.Tabs 
+#     component, an ExplainerHeader(mode='standalone') will insert a dummy 
+#     'tabs' hidden div.
 
-    Even if you don't need or want a positive label selector in your layout, 
-    you still need a 'pos-label' input element in order for the callbacks
-    to be validated. an ExplainerHeader(mode='hidden'), will add a hidden
-    div with both a 'tabs' and a 'pos-label' element.
+#     Even if you don't need or want a positive label selector in your layout, 
+#     you still need a 'pos-label' input element in order for the callbacks
+#     to be validated. an ExplainerHeader(mode='hidden'), will add a hidden
+#     div with both a 'tabs' and a 'pos-label' element.
 
-    For a subcomponent, you don't need to define either a 'pos-label' or
-    a 'tabs' element (these are taken care off by the overall layout),
-    so ExplainerHeader(mode='none'), will simply add a dummy hidden layout.
-    """
-    def __init__(self, explainer, mode="none", title="Explainer Dashboard",
-                    hide_title=False, hide_selector=False, fluid=True):
-        """Insert appropriate header layout into explainable dashboard layout.
+#     For a subcomponent, you don't need to define either a 'pos-label' or
+#     a 'tabs' element (these are taken care off by the overall layout),
+#     so ExplainerHeader(mode='none'), will simply add a dummy hidden layout.
+#     """
+#     def __init__(self, explainer, mode="none", title="Explainer Dashboard",
+#                     hide_title=False, hide_selector=False, fluid=True):
+#         """Insert appropriate header layout into explainable dashboard layout.
 
-        Args:
-            explainer ([type]): [description]
-            mode (str, {'dashboard', 'standalone', 'hidden', 'none', optional): 
-                "dashboard": Display both title and label selector.
-                    Used for ExplainerDashboard.
-                "dashboard_titleonly": Display title and add dummy label selector
-                "standalone": display title and label selector, insert dummy 'tabs'.
-                    Used for ExplainerTabs or standalone single layouts.
-                "hidden": don't display header but insert dummy elements 'pos-label' and 'tabs'. 
-                    Used for InlineExplainer.
-                "none": don't generate any header at all.
-                    Used for sub ExplainerComponents.. Defaults to "none".
-            title (str, optional): Title to display in header. 
-                    Defaults to "Explainer Dashboard".
-            hide_title(bool, optional): Hide the title. Defaults to False.
-            hide_selector(bool, optional): Hide the selector. Defaults to False.
-            fluid(bool, optional): whether to "stretch" (fluid=True) the layout
-                    of the dbc.Container([])
+#         Args:
+#             explainer ([type]): [description]
+#             mode (str, {'dashboard', 'standalone', 'hidden', 'none', optional): 
+#                 "dashboard": Display both title and label selector.
+#                     Used for ExplainerDashboard.
+#                 "dashboard_titleonly": Display title and add dummy label selector
+#                 "standalone": display title and label selector, insert dummy 'tabs'.
+#                     Used for ExplainerTabs or standalone single layouts.
+#                 "hidden": don't display header but insert dummy elements 'pos-label' and 'tabs'. 
+#                     Used for InlineExplainer.
+#                 "none": don't generate any header at all.
+#                     Used for sub ExplainerComponents.. Defaults to "none".
+#             title (str, optional): Title to display in header. 
+#                     Defaults to "Explainer Dashboard".
+#             hide_title(bool, optional): Hide the title. Defaults to False.
+#             hide_selector(bool, optional): Hide the selector. Defaults to False.
+#             fluid(bool, optional): whether to "stretch" (fluid=True) the layout
+#                     of the dbc.Container([])
             
-        """
-        assert mode in ["dashboard", "standalone", "hidden", "none"]
-        self.explainer = explainer
-        self.mode = mode
-        self.title = title
-        self.hide_title, self.hide_selector = hide_title, hide_selector
-        self.fluid = fluid
+#         """
+#         assert mode in ["dashboard", "standalone", "hidden", "none"]
+#         self.explainer = explainer
+#         self.mode = mode
+#         self.title = title
+#         self.hide_title, self.hide_selector = hide_title, hide_selector
+#         self.fluid = fluid
 
-    def layout(self):
-        """html.Div() with elements depending on self.mode:
+#     def layout(self):
+#         """html.Div() with elements depending on self.mode:
 
-        - 'dashboard': title+positive label selector
-        - 'standalone': title+positive label selector + hidden 'tabs' element
-        - 'hidden': hidden 'pos-label' and 'tabs' element
-        - 'none': empty layout
-        """
-        dummy_pos_label = html.Div(
-                [dcc.Input(id="pos-label")], style=dict(display="none"))
-        dummy_tabs = None #html.Div(dcc.Input(id="tabs"), style=dict(display="none"))
+#         - 'dashboard': title+positive label selector
+#         - 'standalone': title+positive label selector + hidden 'tabs' element
+#         - 'hidden': hidden 'pos-label' and 'tabs' element
+#         - 'none': empty layout
+#         """
+#         dummy_pos_label = html.Div(
+#                 [dcc.Input(id="pos-label")], style=dict(display="none"))
+#         dummy_tabs = None #html.Div(dcc.Input(id="tabs"), style=dict(display="none"))
 
-        title_col = make_hideable(
-            dbc.Col([html.H1(self.title)], width='auto'), hide=self.hide_title)
+#         title_col = make_hideable(
+#             dbc.Col([html.H1(self.title)], width='auto'), hide=self.hide_title)
 
 
-        if self.explainer.is_classifier:
-            pos_label_group = make_hideable(
-                dbc.Col([
-                    dbc.Label("Positive class:", html_for="pos-label"),
-                    dcc.Dropdown(
-                        id='pos-label',
-                        options = [{'label': label, 'value': i}
-                                for i, label in enumerate(self.explainer.labels)],
-                        value = self.explainer.pos_label)
-                ], width=2), 
-                hide=self.hide_selector)
-        else:
-            pos_label_group = html.Div(
-                [dcc.Input(id="pos-label")], style=dict(display="none"))
+#         if self.explainer.is_classifier:
+#             pos_label_group = make_hideable(
+#                 dbc.Col([
+#                     dbc.Label("Positive class:", html_for="pos-label"),
+#                     dcc.Dropdown(
+#                         id='pos-label',
+#                         options = [{'label': label, 'value': i}
+#                                 for i, label in enumerate(self.explainer.labels)],
+#                         value = self.explainer.pos_label)
+#                 ], width=2), 
+#                 hide=self.hide_selector)
+#         else:
+#             pos_label_group = html.Div(
+#                 [dcc.Input(id="pos-label")], style=dict(display="none"))
 
-        if self.mode=="dashboard":
-            return dbc.Container([
-                dbc.Row([title_col, pos_label_group],
-                justify="start", align="center")
-            ], fluid=self.fluid)
-        elif self.mode=="standalone":
-            return dbc.Container([
-                dbc.Row([title_col, pos_label_group, dummy_tabs],
-                justify="start", align="center")
-            ], fluid=self.fluid)
-        elif self.mode=="hidden":
-            return html.Div(
-                    dbc.Row([pos_label_group, dummy_tabs])
-            , style=dict(display="none"))
-        elif self.mode=="none":
-            return None
+#         if self.mode=="dashboard":
+#             return dbc.Container([
+#                 dbc.Row([title_col, pos_label_group],
+#                 justify="start", align="center")
+#             ], fluid=self.fluid)
+#         elif self.mode=="standalone":
+#             return dbc.Container([
+#                 dbc.Row([title_col, pos_label_group, dummy_tabs],
+#                 justify="start", align="center")
+#             ], fluid=self.fluid)
+#         elif self.mode=="hidden":
+#             return html.Div(
+#                     dbc.Row([pos_label_group, dummy_tabs])
+#             , style=dict(display="none"))
+#         elif self.mode=="none":
+#             return None
 
 
 class ExplainerComponent(ABC):
@@ -213,42 +211,32 @@ class ExplainerComponent(ABC):
 
     ExplainerComponent makes sure that:
 
-    1. An appropriate header is inserted into .layout()
-    2. Callbacks of subcomponents are registered.
-    3. Lazily calculated dependencies (even of subcomponents) can be calculated.
+    1. Callbacks of subcomponents are registered.
+    2. Lazily calculated dependencies (even of subcomponents) can be calculated.
+    3. Pos labels selector id's of all subcomponents can be calculated. 
     
     Each ExplainerComponent adds a unique uuid name string to all elements, so 
     that there is never a name clash even with multiple ExplanerComponents of 
     the same type in a layout. 
 
     Important:
-        define your layout in _layout() and your callbacks in _register_callbacks()
-        ExplainerComponent will return a layout() consisting of both
-        header and _layout(), and register callbacks of subcomponents in addition
+        define your callbacks in _register_callbacks() and
+        ExplainerComponent will register callbacks of subcomponents in addition
         to _register_callbacks() when calling register_callbacks()
     """
-    def __init__(self, explainer, title=None, header_mode="none", name=None):
+    def __init__(self, explainer, title=None, name=None):
         """initialize the ExplainerComponent
 
         Args:
             explainer (Explainer): explainer object constructed with e.g.
                         ClassifierExplainer() or RegressionExplainer()
             title (str, optional): Title of tab or page. Defaults to None.
-            header_mode (str,{"standalone", "hidden" or "none"}, optional):
-                        determines what kind of ExplainerHeader to insert into
-                        the layout. For a component that forms the entire page, 
-                        use 'standalone'. If you don't want to see the title and
-                        pos label selector, use 'hidden'. If this is a subcomponent
-                        of a larger page, use 'none'.
-                        Defaults to "none".
             name (str, optional): unique name to add to Component elements. 
                         If None then random uuid is generated to make sure 
                         it's unique. Defaults to None.
-
         """
         self.explainer = explainer
         self.title = title
-        self.header = ExplainerHeader(explainer, title=title, mode=header_mode)
         self.name = name
         if self.name is None:
             self.name = shortuuid.ShortUUID().random(length=10)
@@ -327,20 +315,10 @@ class ExplainerComponent(ABC):
                 ValueError(f"Failed to generate dependency '{dep}': "
                     "Failed to calculate or retrieve explainer property explainer.{dep}...")
 
-    def _layout(self):
+    def layout(self):
         """layout to be defined by the particular ExplainerComponent instance.
         All element id's should append +self.name to make sure they are unique."""
         return None
-
-    def layout(self):
-        "returns a combination of the header and _layout()"
-        if not hasattr(self, 'header'):
-            return self._layout()
-        else:
-            return html.Div([
-                self.header.layout(),
-                self._layout()
-            ])
 
     def _register_callbacks(self, app):
         """register callbacks specific to this ExplainerComponent"""
@@ -359,9 +337,10 @@ class ExplainerComponent(ABC):
 
 class PosLabelSelector(ExplainerComponent):
     def __init__(self, explainer, title='Pos Label Selector',
-                     header_mode="none", name=None):
+                     header_mode="none", name=None,
+                     pos_label=None):
         super().__init__(explainer, title, header_mode, name)
-        
+        self.pos_label = pos_label if pos_label is not None else explainer.pos_label
     def layout(self):
         if self.explainer.is_classifier:
             return html.Div([
@@ -370,169 +349,7 @@ class PosLabelSelector(ExplainerComponent):
                                 id='pos-label-'+self.name,
                                 options = [{'label': label, 'value': i}
                                         for i, label in enumerate(self.explainer.labels)],
-                                value = self.explainer.pos_label)
+                                value = self.pos_label)
                             ])
         else:
             return html.Div([dcc.Input(id="pos-label-"+self.name)], style=dict(display="none"))
-
-
-
-def instantiate_component(component, explainer, header_mode="none", **kwargs):
-    """Returns an instantiated ExplainerComponent.
-    
-    If the component input is just a class definition, instantiate it with
-    explainer, header_mode and kwargs.
-    If it is already an ExplainerComponent instance, set header_mode and return it.
-    If it is any other instance with layout and regiuster_components methods,
-    then add a name property and return it. 
-
-    Args:
-        component ([type]): Either a class definition or instance
-        explainer ([type]): An Explainer object that will be used to instantiate class definitions
-        header_mode (str, optional): Header_mode to assign to component. Defaults to "none".
-        kwargs: kwargs will be passed on to the instance
-
-    Raises:
-        ValueError: if component is not a subclass or instance of ExplainerComponent,
-                or is an instance without layout and register_callbacks methods
-
-    Returns:
-        [type]: instantiated component
-    """
-
-    if inspect.isclass(component) and issubclass(component, ExplainerComponent):
-        return component(explainer, header_mode=header_mode, **kwargs)
-    elif isinstance(component, ExplainerComponent):
-        component.header.mode = header_mode
-        return component
-    elif (not inspect.isclass(component)
-          and hasattr(component, "layout")):
-        if not (hasattr(component, "name") and isinstance(component.name, str)):
-            try:
-                component_name  = component.__name__
-            except:
-                component_name = shortuuid.ShortUUID().random(length=10)
-            print(f"Warning: setting {component}.name to {component_name}")
-            component.name = component_name
-        if not hasattr(component, "title"):
-            print(f"Warning: setting {component}.title to 'CustomTab'")
-            component.title = "CustomTab"
-        try:
-            component.header.mode = header_mode
-        except AttributeError:
-            print(f"Warning couldn't find a header to set mode to {header_mode}")
-        return component
-    else:
-        raise ValueError(f"{component} is not a valid component...")
-
-
-class ExplainerTabsLayout:
-    def __init__(self, explainer, tabs,
-                 title='Model Explainer',
-                 hide_title=False,
-                 hide_selector=False,
-                 fluid=True,
-                 **kwargs):
-        """Generates a multi tab layout from a a list of ExplainerComponents.
-        If the component is a class definition, it gets instantiated first. If 
-        the component is not derived from an ExplainerComponent, then attempt
-        with duck typing to nevertheless instantiate a layout.
-
-        Args:
-            explainer ([type]): explainer
-            tabs (list[ExplainerComponent class or instance]): list of
-                ExplainerComponent class definitions or instances.
-            title (str, optional): [description]. Defaults to 'Model Explainer'.
-            hide_title (bool, optional): Hide the title. Defaults to False.
-            hide_selector (bool, optional): Hide the positive label selector. 
-                        Defaults to False.
-            fluid (bool, optional): Stretch layout to fill space. Defaults to False.
-        """
-        self.fluid = fluid
-
-        self.header = ExplainerHeader(explainer, "dashboard", 
-                                    hide_title=hide_title, 
-                                    hide_selector=hide_selector,
-                                    fluid=fluid)
-        self.tabs  = [instantiate_component(tab, explainer, "none", **kwargs) for tab in tabs]
-        assert len(self.tabs) > 0, 'When passing a list to tabs, need to pass at least one valid tab!'
-        
-    def layout(self):
-        """returns a multitab layout plus ExplainerHeader"""
-        return dbc.Container([
-            self.header.layout(),
-            dcc.Tabs(id="tabs", value=self.tabs[0].name, 
-                        children=[dcc.Tab(label=tab.title, id=tab.name, value=tab.name,
-                                        children=tab.layout()) for tab in self.tabs]),
-        ], fluid=self.fluid)
-
-    def register_callbacks(self, app):
-        """Registers callbacks for all tabs"""
-        for tab in self.tabs:
-            try:
-                tab.register_callbacks(app)
-            except AttributeError:
-                print(f"Warning: {tab} does not have a register_callbacks method!")
-
-    def calculate_dependencies(self):
-        """Calculates dependencies for all tabs"""
-        for tab in self.tabs:
-            try:
-                tab.calculate_dependencies()
-            except AttributeError:
-                print(f"Warning: {tab} does not have a calculate_dependencies method!")
-
-
-class ExplainerPageLayout(ExplainerComponent):
-    def __init__(self, explainer, component,
-                 title='Model Explainer',
-                 hide_title=False,
-                 hide_selector=False,
-                 fluid=False,
-                 **kwargs):
-        """Generates a single page layout from a single ExplainerComponent.
-        If the component is a class definition, it gets instantiated. 
-
-        If the component is not derived from an ExplainerComponent, then tries
-        with duck typing to nevertheless instantiate a layout.
-
-
-        Args:
-            explainer ([type]): explainer
-            component (ExplainerComponent class or instance): ExplainerComponent 
-                        class definition or instance.
-            title (str, optional): [description]. Defaults to 'Model Explainer'.
-            hide_title (bool, optional): Hide the title. Defaults to False.
-            hide_selector (bool, optional): Hide the positive label selector. 
-                        Defaults to False.
-            fluid (bool, optional): Stretch layout to fill space. Defaults to False.
-        """
-        
-        self.fluid = fluid
-        self.header = ExplainerHeader(explainer, "standalone", 
-                                    hide_title=hide_title, 
-                                    hide_selector=hide_selector,
-                                    fluid=fluid)
-        self.page  = instantiate_component(component, explainer, "none", **kwargs)
-        
-    def layout(self):
-        """returns single page layout with an ExplainerHeader"""
-        return dbc.Container([
-            self.header.layout(),
-            self.page.layout()
-        ], fluid=self.fluid)
-
-    def register_callbacks(self, app):
-        """Register callbacks of page"""
-        try:
-            self.page.register_callbacks(app)
-        except AttributeError:
-            print(f"Warning: {self.page} does not have a register_callbacks method!")
-
-    def calculate_dependencies(self):
-        """Calculate dependencies of page"""
-        try:
-            self.page.calculate_dependencies()
-        except AttributeError:
-            print(f"Warning: {self.page} does not have a calculate_dependencies method!")
-
